Style & Edit: 게시물 alert 및 homepost를 homeimgpost와 합침
