+++ src/App.jsx
@@ -28,7 +28,7 @@ function App() {
                 )}
             />
             <Route path='/profilesetting' component={() => <ProfileSettingPage />} />
-            <Route path='/home' component={() => <HomePage />} />
+            <Route path='/home' component={() => <HomePage localToken={localToken} />} />
             <Route path='/search' component={() => <SearchPage />} />
 
             <Route
